Move `db/graphql` folder content to `graphql` – backend
