+++ backend/src/middleware/slugifyMiddleware.spec.js
@@ -2,9 +2,9 @@ import { getNeode, getDriver } from '../db/neo4j'
 import createServer from '../server'
 import { createTestClient } from 'apollo-server-testing'
 import Factory, { cleanDatabase } from '../db/factories'
-import { createGroupMutation, updateGroupMutation } from '../db/graphql/groups'
-import { createPostMutation } from '../db/graphql/posts'
-import { signupVerificationMutation } from '../db/graphql/authentications'
+import { createGroupMutation, updateGroupMutation } from '../graphql/groups'
+import { createPostMutation } from '../graphql/posts'
+import { signupVerificationMutation } from '../graphql/authentications'
 
 let authenticatedUser
 let variables
