+++ backend/src/schema/resolvers/user_management.spec.js
@@ -2,7 +2,7 @@ import jwt from 'jsonwebtoken'
 import CONFIG from './../../config'
 import Factory, { cleanDatabase } from '../../db/factories'
 import { gql } from '../../helpers/jest'
-import { loginMutation } from '../../db/graphql/userManagement'
+import { loginMutation } from '../../graphql/userManagement'
 import { createTestClient } from 'apollo-server-testing'
 import createServer, { context } from '../../server'
 import encode from '../../jwt/encode'
